requirements.txt
PySide6
pyyaml
ansible-runner